py/requirements.txt
grpcio-tools~=1.47